Add databricks test

+++ examples/dbt/jaffle_shop/macros/drop_table.sql
@@ -1,5 +1,5 @@
 {%- macro drop_table(table_name, conn_type) -%}
-    {% if conn_type != 'bigquery' %}
+    {% if conn_type != 'bigquery' and conn_type != 'databricks' %}
       {%- set drop_query -%}
           DROP TABLE IF EXISTS {{ target.schema }}.{{ table_name }} CASCADE
       {%- endset -%}
